Creating the Texture classes. Seriously thinking in not add texture to the shapes painters.. Textures are resources that need to be managed by shaders who load them to the GPU as they are needed by the shapes in the DrawTree.  Shapes need to know only how their trianles map to a texture.
Yes, the mapping function is the glue. For the shapes a Painter is just a mapping to a texture.
On the side of code usability, programmers will always liek to associate shapes with textures. So the link between shapes and textures should be defined and exported to programmers. Programmers should be able to change the texture of a shape by accessing the shape. That will cause changes in the DrawTree creating nodes to load this new Texture in the GPU.

Yes, Usability is forcing me to create textures as part of the shape. This associations must cause
changes in the Shader DrawTree. These changes must be triggered behind the scenes. Programmers should never have access to the DrawTree, unless they are programming the shaders themselves. Need to design the behinh the scenes mechanism