Adding TextureSlots to HLSLProgram.cs To be used by the shaders. I am replacing the effects framework.
Seriously Thinking to create a  Texture manager resource in the IDxContext level to take care of clean unused textures.

Also the Textures must just have a FileInfo containing the image file.
It will be a job of the framework load from disk when it is used by the shaders and disallocate when it is not. In other words create a cache system for HardDisk and GPU memory.