Creating a ShaderBase class that implements the IObservable interface. The ShaderBase has the DrawTree object now and it overlays its change event. When you attach an observer to ShaderBase you also attaches an observer to drawtree.
Notes:
I don't know if this is a good idea. Now the childs of ChildBase will have to use the DrawTree of its mother, and also remember to call the dispose of its mother. Also adding members to mother base (state) forces the childs to understand the mother state.  Prefering Inheritance over Agregation generally result in buggy frameworks.