Ok, Inheritance stinks a lot. The developer of the child will have always to check the mother class to figure it out what it is doing with their members. If it is disposing them or relying in the childs to do that, the  developer have to be smart enough to use the mothers draw tree instead of defining its own otherwise the observer will not work. Besides that, the developer will  see the mother as one more layer to implement a shader. One more obstacle.
It is better to only require the shaders to implement interfaces. The developer is more confident in this case that if he implements the interface they will be able to attach a shader in the framework.

Implementation Detail: Changing inheritance by aggregation.

IShader has a property IObservable that will be used by the class clients to attach observers. The implementors have to create this IObservable implementation and expose it.  To help them, I created an ShaderObservableDock class that implments the IObservable. This class is basically a relay. It receives an IObservable object  during its construction. When an observer attaches itself,  the ShaderObservabledock, attaches it to the IObservable defined in its constructor.

Shaders usually pass to ShaderObservableDock constructor its own DrawTree, and therefore their observers will always be notifies when the DrawTree changes DrawThis is used by  the s