Errors sometimes are good, if you are looking in the right way. Just got a bug that indicates a design flaw.
The DxContext has an addshader method and shaders are required to add itself on that context. Not this is a problem. Because the shaders add themselves to dxcontext when they are constructing themselves. So DxControl.AddShader() will receive a shader that has incomplete state. Bad design, There is two way dependency between dxcontext and shader. This must end. Shader uses dxcontext and period, not the other way around.

AddShader is now a functionality outside IDxContext. It is not part of the interface.
The addition of the shader happens outside and elsewhere. In the case of the current implementation of IDxContext, DxControl will have an AddShader option to be used by the clients to hookup the shaders. Clients create the shaders with an IDxContext provided by DxControl. (In this case DxControl itself implements IDxContext)  and then after the shader creation the clients add the shader to the DxControl.

Thinking to Make DxControl to give an implementation of IDxContext instead of making it implement the IDxContext itself.  On responsability principle. DxControl  should be a wrapper of a windows form that is able to give IDxContext, not be able to be used as an IDxContext.